refactor: centralize session management
